+++ aiprocess/build.gradle
@@ -1,25 +1,22 @@
-bootJar { enabled = true }
+bootJar { enabled = false }
 
-jar { enabled = false }
+jar { enabled = true }
+
+version="2.0.0"
 
 dependencies {
     implementation project(':common')
     implementation project(":common-event")
-    implementation project(":auth")
 
-    implementation 'org.springframework.boot:spring-boot-starter-security'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation("org.springframework.boot:spring-boot-starter-security")
 
     // kafka
     implementation 'org.springframework.cloud:spring-cloud-stream'
     implementation 'org.springframework.cloud:spring-cloud-stream-binder-kafka'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testImplementation 'org.springframework.cloud:spring-cloud-stream-test-binder'
-
-    // jpa
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    runtimeOnly 'com.mysql:mysql-connector-j'
 
     // avro
     implementation('org.apache.avro:avro:1.11.4')
