+++ aiprocess/src/main/java/org/newshabit/app/aiprocess/domain/dto/AiProcessedNews.java
@@ -1,0 +1,12 @@
+package org.newshabit.app.aiprocess.domain.dto;
+
+public record AiProcessedNews(
+	String title,
+	String who,
+	String when,
+	String where,
+	String what,
+	String why,
+	String how,
+	String keyword,
+	String summary) {}
