+++ aiprocess/src/main/java/org/newshabit/app/aiprocess/domain/dto/GeminiResponse.java
@@ -1,0 +1,17 @@
+package org.newshabit.app.aiprocess.domain.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import java.util.List;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public record GeminiResponse(List<Candidate> candidates) {
+
+	@JsonIgnoreProperties(ignoreUnknown = true)
+	public record Candidate(Content content, String finishReason, double avgLogProbs) {}
+
+	@JsonIgnoreProperties(ignoreUnknown = true)
+	public record Content(List<Part> parts, String role) {}
+
+	@JsonIgnoreProperties(ignoreUnknown = true)
+	public record Part(String text) {}
+}
