+++ aiprocess/src/main/resources/application-ai.yml
