+++ aiprocess/src/main/resources/application-app.yml
@@ -1,0 +1,2 @@
+server:
+  port: 8081
