+++ aiprocess/src/main/resources/application-spring-cloud-stream.yml
@@ -1,0 +1,44 @@
+spring:
+  cloud:
+    function:
+      definition: refineNews;
+    stream:
+      bindings:
+        refineNews-in-0:
+          destination: ${CRAWL_TOPIC}
+          content-type: application/avro
+          group: refine-group
+          consumer:
+            partitioned: true
+            maxAttempts: 3
+            back-off-initial-interval: 1000
+            concurrency: 1
+            use-native-decoding: true
+        refineNews-out-0:
+          destination: ${REFINED_NEWS_TOPIC}
+          content-type: application/avro
+          group: refine-group
+          producer:
+            use-native-encoding: true
+      kafka:
+        binder:
+          brokers: localhost:29092
+          auto-create-topics: false
+        bindings:
+          refineNews-in-0:
+            consumer:
+              configuration:
+                key.deserializer: io.confluent.kafka.serializers.KafkaAvroDeserializer
+                value.deserializer: io.confluent.kafka.serializers.KafkaAvroDeserializer
+                schema.registry.url: ${SCHEMA_REGISTRY_URL}
+                linger.ms: 1000
+                specific.avro.reader: true
+          refineNews-out-0:
+            producer:
+              configuration:
+                key.serializer: io.confluent.kafka.serializers.KafkaAvroSerializer
+                value.serializer: io.confluent.kafka.serializers.KafkaAvroSerializer
+                schema.registry.url: ${SCHEMA_REGISTRY_URL}
+                linger.ms: 1000
+                auto.register.schemas: true
+                specific.avro.reader: true
