+++ auth/docker-compose.yml
@@ -1,0 +1,26 @@
+services:
+  newshabit-auth:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    image: ${NEWSHABIT_AUTH_IMAGE_NAME}:${NEWSHABIT_AUTH_IMAGE_TAG}
+    container_name: newshabit-auth
+    restart: always
+    ports:
+      - "8082:8080"
+    environment:
+      JAVA_OPTS: "-Xms128m -Xmx512m -XX:+UseG1GC"
+      SPRING_PROFILES_ACTIVE: "app,auth"
+      JWT_SECRET_KEY: ${JWT_SECRET_KEY}
+      ACCESS_TOKEN_VALID_TIME: ${ACCESS_TOKEN_VALID_TIME}
+      REFRESH_TOKEN_VALID_TIME: ${REFRESH_TOKEN_VALID_TIME}
+
+    volumes:
+      - /data/auth/log:/app/log
+      - /usr/share/zoneinfo/Asia/Seoul:/etc/localtime:ro
+    networks:
+      - newshabit-2
+
+networks:
+  newshabit-2:
+    driver: bridge
