--- docker-compose.newshabit-3.yml
@@ -1,47 +1,0 @@
-services:
-  newshabit-auth:
-    build:
-      context: .
-      dockerfile: auth/Dockerfile
-    image: ${NEWSHABIT_AUTH_IMAGE_NAME}:${NEWSHABIT_AUTH_IMAGE_TAG}
-    container_name: newshabit-auth
-    restart: always
-    volumes:
-      - /data/auth/log:/app/log
-      - /usr/share/zoneinfo/Asia/Seoul:/etc/localtime:ro
-    ports:
-      - "8081:8080"
-    environment:
-      JAVA_OPTS: "-Xms128m -Xmx512m -XX:+UseG1GC"
-      SPRING_PROFILES_ACTIVE: "app,auth"
-      JWT_SECRET_KEY: ${JWT_SECRET_KEY}
-    networks:
-      - newshabit-3
-
-  newshabit-user:
-    build:
-      context: .
-      dockerfile: user/Dockerfile
-    image: ${NEWSHABIT_USER_IMAGE_NAME}:${NEWSHABIT_USER_IMAGE_TAG}
-    container_name: newshabit-user
-    restart: always
-    volumes:
-      - /data/user/log:/app/log
-      - /usr/share/zoneinfo/Asia/Seoul:/etc/localtime:ro
-    ports:
-      - "8082:8080"
-    environment:
-      JAVA_OPTS: "-Xms128m -Xmx512m -XX:+UseG1GC"
-      SPRING_PROFILES_ACTIVE: "app,auth,jdbc"
-      HEADLINE_URI: ${HEADLINE_URI}
-      USER_AGENT: ${USER_AGENT}
-      JWT_SECRET_KEY: ${JWT_SECRET_KEY}
-      RDS_URL: ${RDS_URL}
-      RDS_USERNAME: ${RDS_USERNAME}
-      RDS_PASSWORD: ${RDS_PASSWORD}
-    networks:
-      - newshabit-3
-
-networks:
-  newshabit-3:
-    driver: bridge
