+++ news/src/main/java/org/newshabit/app/news/application/port/RefineNewsUseCase.java
@@ -1,0 +1,8 @@
+package org.newshabit.app.news.application.port;
+
+
+import org.newshabit.app.news.domain.entity.RefinedNewsEntity;
+
+public interface RefineNewsUseCase {
+	void sinkRefinedNews(RefinedNewsEntity refinedNewsEntity);
+}
