+++ news/src/main/java/org/newshabit/app/news/application/port/RefinedNewsRepositoryOutputPort.java
@@ -1,0 +1,9 @@
+package org.newshabit.app.news.application.port;
+
+
+import org.newshabit.app.news.domain.entity.RefinedNewsEntity;
+
+public interface RefinedNewsRepositoryOutputPort {
+	boolean existsByOriginalUrl(String url);
+	RefinedNewsEntity save(RefinedNewsEntity refinedNewsEntity);
+}
