+++ news/src/main/java/org/newshabit/app/news/infrastructure/repository/RefinedNewsRepository.java
@@ -1,0 +1,12 @@
+package org.newshabit.app.news.infrastructure.repository;
+
+
+import org.newshabit.app.news.domain.entity.RefinedNewsEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RefinedNewsRepository extends JpaRepository<RefinedNewsEntity, Long> {
+
+	boolean existsByOriginalUrl(String url);
+}
