+++ news/src/main/resources/application-app.yml
