+++ news/src/main/resources/application-auth.yml
