+++ news/src/main/resources/application-spring-cloud-stream.yml
@@ -1,0 +1,31 @@
+spring:
+  cloud:
+    function:
+      definition: sinkRefinedNews;
+    stream:
+      bindings:
+        sinkRefinedNews-in-0:
+          destination: ${REFINED_NEWS_TOPIC}
+          content-type: application/avro
+          group: sink-group
+          consumer:
+            partitioned: true
+            maxAttempts: 3
+            back-off-initial-interval: 1000
+            concurrency: 3
+            use-native-decoding: true
+      kafka:
+        binder:
+          brokers: localhost:29092
+          auto-create-topics: false
+        bindings:
+          sinkRefinedNews-in-0:
+            consumer:
+              reset-offsets: true
+              start-offset: latest
+              configuration:
+                key.deserializer: io.confluent.kafka.serializers.KafkaAvroDeserializer
+                value.deserializer: io.confluent.kafka.serializers.KafkaAvroDeserializer
+                schema.registry.url: ${SCHEMA_REGISTRY_URL}
+                linger.ms: 1000
+                specific.avro.reader: true
