--- pubsub/build.gradle
@@ -1,29 +1,0 @@
-bootJar { enabled = false }
-
-jar { enabled = true }
-
-version="2.0.0"
-
-dependencies {
-    implementation project(':common')
-    implementation project(":common-event")
-    implementation project(":auth")
-
-    implementation 'org.springframework.boot:spring-boot-starter-validation'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation("org.springframework.boot:spring-boot-starter-security")
-
-    // kafka
-    implementation 'org.springframework.cloud:spring-cloud-stream'
-    implementation 'org.springframework.cloud:spring-cloud-stream-binder-kafka'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation 'org.springframework.cloud:spring-cloud-stream-test-binder'
-
-    // jpa
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    runtimeOnly 'com.mysql:mysql-connector-j'
-
-    // avro
-    implementation('org.apache.avro:avro:1.11.4')
-    implementation("io.confluent:kafka-avro-serializer:7.8.0")
-}
