--- pubsub/src/main/java/org/newshabit/app/pubsub/application/port/RefinedNewsRepositoryOutputPort.java
@@ -1,8 +1,0 @@
-package org.newshabit.app.pubsub.application.port;
-
-import org.newshabit.app.pubsub.domain.entity.RefinedNewsEntity;
-
-public interface RefinedNewsRepositoryOutputPort {
-	boolean existsByOriginalUrl(String url);
-	RefinedNewsEntity save(RefinedNewsEntity refinedNewsEntity);
-}
