+++ settings.gradle
@@ -2,7 +2,7 @@ rootProject.name = 'app'
 include 'crawl'
 include 'news'
 include 'common'
-include 'pubsub'
+include 'aiprocess'
 include 'user'
 include 'common-event'
 include 'auth'
