+++ user/docker-compose.yml
@@ -1,0 +1,28 @@
+services:
+  newshabit-user:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    image: ${NEWSHABIT_USER_IMAGE_NAME}:${NEWSHABIT_USER_IMAGE_TAG}
+    container_name: newshabit-user
+    restart: always
+    volumes:
+      - /data/user/log:/app/log
+      - /usr/share/zoneinfo/Asia/Seoul:/etc/localtime:ro
+    ports:
+      - "8081:8080"
+    environment:
+      JAVA_OPTS: "-Xms128m -Xmx512m -XX:+UseG1GC"
+      SPRING_PROFILES_ACTIVE: "app,auth,jdbc"
+      JWT_SECRET_KEY: ${JWT_SECRET_KEY}
+      ACCESS_TOKEN_VALID_TIME: ${ACCESS_TOKEN_VALID_TIME}
+      REFRESH_TOKEN_VALID_TIME: ${REFRESH_TOKEN_VALID_TIME}
+      RDS_URL: ${RDS_URL}
+      RDS_USERNAME: ${RDS_USERNAME}
+      RDS_PASSWORD: ${RDS_PASSWORD}
+    networks:
+      - newshabit-2
+
+networks:
+  newshabit-2:
+    driver: bridge
